src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/pseudo-not-expected.txt

CONSOLE ERROR: line 14: Uncaught TypeError: host.createShadowRoot is not a function
Test the :not pseudo selector in

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
